scripts: referencegen: Date function: Remove delimiter method
Using the comma as a delimiter is an obstruction on the user as it is not common to input a comma in the middle of a date and therefore may forget to. This method should also make it feasible to copy and paste in the date from original source. May use this method to grab author names