scripts: referencegen: Author function: Remove delimiter method
Removing delimiter method, the script can use the position method to verify if the user has inputted the corrected number of names corresponding to the number authors. This was done by placing a while loop until quota was matched.
Misc: Improved colour display of terminal, improved input for authors to support easier copy and pasting